fragment supplémentaires

+++ src/main/java/com/ipiecoles/communes/web/controller/CommuneController.java
@@ -52,11 +52,14 @@ public class CommuneController {
     @PostMapping(value ="/communes", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
     public String saveNewCommune(
             Commune commune,
+            RedirectAttributes attributes,
             final ModelMap model)
     {
         // Ajouter un certain nombre de contrôles...
         commune = communeRepository.save(commune);
         model.put("commune", commune);
+        attributes.addFlashAttribute("type", "success");
+        attributes.addFlashAttribute("message", "Enregistrement de la nouvelle commune effectué avec succès.");
         return "redirect:/communes/" + commune.getCodeInsee();
     }
 
@@ -79,9 +82,18 @@ public class CommuneController {
 
     @GetMapping("/communes/{codeInsee}/delete")
     public String deleteCommune(
-            @PathVariable String codeInsee)
+            @PathVariable String codeInsee,
+            RedirectAttributes attributes)
     {
-        communeRepository.deleteById(codeInsee);
+        Optional<Commune> commune = communeRepository.findById(codeInsee);
+        if(commune.isPresent())
+        {
+            communeRepository.deleteById(codeInsee);
+        } else {
+            throw new EntityNotFoundException("Impossible de trouver la commune ayant pour code INSEE: " + codeInsee);
+        }
+        attributes.addFlashAttribute("type", "success");
+        attributes.addFlashAttribute("message", "Suppression de la commune effectué avec succès.");
         return "redirect:/";
     }
